fix(php): path selection, path-request throttling, and stale-state cleanup
Fixes a set of routing and storage bugs found while tracing browser-to-browser
and distro traffic across the relay chain. Behaviour is matched to the Python
reference (RNS/Transport.py) where an equivalent exists.

Path selection (request_control_plane_trait.php)
- A strictly shorter path is always better routing information, but the
  random-blob replay guard was rejecting it: the same announce commonly
  arrives via both a short direct peer and a longer transit route, and
  whichever landed first claimed the blob. Observed in production as
  retichat.com holding a 6-hop gateway path to a selectiv-local destination
  instead of the 2-hop direct peer, which dropped direct messages at the
  gateway.
- Announce emission time is untrusted network input and is only used to
  decide freshness, so a future-dated value would pin the comparison
  baseline forever and make the destination permanently unrefreshable.
  Anything beyond a small clock-skew allowance is now treated as unusable.

Path requests (request_path_state_trait.php)
- Adds the minimum interval between automated path requests and the pending
  discovery window (Transport.py:78,81). Python gates a new search on
  should_search_for_unknown; without it every retry carried a fresh random
  tag, so tag dedup never fired and each retry re-flooded every local client.
- A path entry whose cached announce has been purged is now treated as no
  path at all, so the discovery branches ask the network instead of silently
  answering nothing until the entry expires.

Local delivery (request_relay_routing_trait.php)
- A local destination is only deliverable while the interface it was
  registered on is still reachable. Local delivery is decided before the
  path table is consulted, so stale rows left behind by a closed browser
  black-holed traffic for destinations that had since moved elsewhere.

Storage and maintenance (request_maintenance_trait.php, schema, config)
- Adds packet_storage_max_bytes (default 300 MB) and a combined storage cap
  so payload tables cannot grow without bound.
- New cleanup phase for orphaned local destinations belonging to stale
  interfaces.

Tests: adds path_selection, path_request_throttle, stale_interface and
packet_storage_cap suites; extends gateway_path_request. All green except
the pre-existing hops_test failure (remaining_hops link-transport), which
fails identically on HEAD and is unrelated to these changes.

+++ php/src/lib/database.php
@@ -247,7 +247,7 @@ final class Database
         $stmt->execute();
         $row = $stmt->fetch(PDO::FETCH_ASSOC);
 
-        return ($row['acquired'] ?? 0) === 1;
+        return (int) ($row['acquired'] ?? 0) === 1;
     }
 
     /**
@@ -266,7 +266,7 @@ final class Database
         $stmt->execute();
         $row = $stmt->fetch(PDO::FETCH_ASSOC);
 
-        return ($row['released'] ?? 0) === 1;
+        return (int) ($row['released'] ?? 0) === 1;
     }
 
     /**
